finished project

+++ src/main/java/pl/connectis/restaurant/RestaurantApp/controller/ClientController.java
@@ -3,7 +3,11 @@ package pl.connectis.restaurant.RestaurantApp.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import pl.connectis.restaurant.RestaurantApp.dto.ClientDTO;
+import pl.connectis.restaurant.RestaurantApp.exception.NotFoundException;
+import pl.connectis.restaurant.RestaurantApp.model.Client;
 import pl.connectis.restaurant.RestaurantApp.service.ClientService;
+
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/client")
@@ -17,7 +21,7 @@ public class ClientController {
     }
 
     @PostMapping(path = "/add")
-    public Long addClient(@RequestBody ClientDTO clientDTO){
+    public Long addClient(@RequestBody ClientDTO clientDTO) {
         Long getId = clientService.addClient(
                 clientDTO.getName(),
                 clientDTO.getSurname(),
@@ -26,21 +30,34 @@ public class ClientController {
         return getId;
     }
 
-    @GetMapping(path = "/get/{id}")
-    public ClientDTO getClient(@PathVariable("id") Long id){
-        ClientDTO clientDTO = new ClientDTO(clientService.getClient(id).get());
-        return clientDTO;
+    @GetMapping(path = "/{id}")
+    public ClientDTO getClient(@PathVariable("id") Long id) {
+        Optional<Client> clientOptional = clientService.getClient(id);
+
+        if (clientOptional.isPresent()) {
+            ClientDTO clientDTO = clientService.toDTO(clientOptional.get());
+            return clientDTO;
+        }
+        throw new NotFoundException();
     }
 
-    @PutMapping(path = "/update/{id}")
-    public String updateClient(@PathVariable("id") Long id,  @RequestParam Double discount){
-        clientService.updateClientDiscount(id, discount);
-        return "Updated Client";
+    @PostMapping(path = "/update/{id}")
+    public String updateClient(@PathVariable("id") Long id, @RequestParam Double discount) {
+        Optional<Client> clientOptional = clientService.getClient(id);
+        if (clientOptional.isPresent()) {
+            clientService.updateClientDiscount(id, discount);
+            return "Updated Client";
+        }
+        throw new NotFoundException();
     }
 
-    @DeleteMapping(path = "/delete/{id}")
-    public String deleteClient(@PathVariable("id") Long id){
-        clientService.removeClient(id);
-        return "Removed client";
+    @DeleteMapping(path = "/{id}")
+    public String deleteClient(@PathVariable("id") Long id) {
+        Optional<Client> clientOptional = clientService.getClient(id);
+        if (clientOptional.isPresent()) {
+            clientService.removeClient(id);
+            return "Removed client";
+        }
+        throw new NotFoundException();
     }
 }
