+++ src/main/java/pl/connectis/restaurant/RestaurantApp/controller/EmplyeeController.java
@@ -3,6 +3,7 @@ package pl.connectis.restaurant.RestaurantApp.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import pl.connectis.restaurant.RestaurantApp.dto.EmployeeDTO;
+import pl.connectis.restaurant.RestaurantApp.exception.NotFoundException;
 import pl.connectis.restaurant.RestaurantApp.model.Employee;
 import pl.connectis.restaurant.RestaurantApp.service.EmployeeService;
 
@@ -19,7 +20,7 @@ public class EmplyeeController {
         this.employeeService = employeeService;
     }
 
-    @PostMapping("/add/owner")
+    @PostMapping("/owner")
     public Long addBoss(@RequestBody EmployeeDTO employeeDTO) {
         Long getId = employeeService.addBoss(
                 employeeDTO.getName(),
@@ -31,7 +32,7 @@ public class EmplyeeController {
         return getId;
     }
 
-    @PostMapping("/add/employee")
+    @PostMapping("/employee")
     public Long addEmployee(@RequestBody EmployeeDTO employeeDTO) {
         Long getId = employeeService.addEmployee(
                 employeeDTO.getName(),
@@ -44,21 +45,34 @@ public class EmplyeeController {
         return getId;
     }
 
-    @GetMapping("/get/{id}")
-    public EmployeeDTO getEmployee(@PathVariable("id") Long id){
+    @GetMapping("/{id}")
+    public EmployeeDTO getEmployee(@PathVariable("id") Long id) {
         Optional<Employee> employeeOptional = employeeService.getEmployee(id);
-        return new EmployeeDTO(employeeOptional.get());
+        if (employeeOptional.isPresent()) {
+            Employee employee = employeeOptional.get();
+            EmployeeDTO employeeDTO = employeeService.toDTO(employee);
+            return employeeDTO;
+        }
+        throw new NotFoundException();
     }
 
-    @PutMapping("/update/{id}")
-    public String updateEmployee(@PathVariable("id") Long id, @RequestParam Long salary){
-        employeeService.updateEmplyee(id, salary);
-        return "emloyee updated";
+    @PostMapping("/update/{id}")
+    public String updateEmployee(@PathVariable("id") Long id, @RequestParam Long salary) {
+        Optional<Employee> employeeOptional = employeeService.getEmployee(id);
+        if (employeeOptional.isPresent()) {
+            employeeService.updateEmplyee(id, salary);
+            return "emloyee updated";
+        }
+        throw new NotFoundException();
     }
 
-    @DeleteMapping("/delete/{id}")
-    public String deleteEmployee(@PathVariable("id") Long id){
-        employeeService.removeEmployee(id);
-        return "employee removed";
+    @DeleteMapping("/{id}")
+    public String deleteEmployee(@PathVariable("id") Long id) {
+        Optional<Employee> employeeOptional = employeeService.getEmployee(id);
+        if (employeeOptional.isPresent()) {
+            employeeService.removeEmployee(id);
+            return "employee removed";
+        }
+        throw new NotFoundException();
     }
 }
