+++ src/main/java/pl/connectis/restaurant/RestaurantApp/controller/StoredItemController.java
@@ -2,6 +2,7 @@ package pl.connectis.restaurant.RestaurantApp.controller;
 
 import org.springframework.web.bind.annotation.*;
 import pl.connectis.restaurant.RestaurantApp.dto.StoredItemDTO;
+import pl.connectis.restaurant.RestaurantApp.exception.NotFoundException;
 import pl.connectis.restaurant.RestaurantApp.model.StoredItem;
 import pl.connectis.restaurant.RestaurantApp.service.StoredItemService;
 
@@ -19,7 +20,7 @@ public class StoredItemController {
     }
 
     @PostMapping("/add")
-    public Long addItem(@RequestBody StoredItemDTO storedItemDTO ){
+    public Long addItem(@RequestBody StoredItemDTO storedItemDTO) {
         Long getId = storedItemService.addItem(
                 storedItemDTO.getName(),
                 storedItemDTO.getQuantity(),
@@ -28,27 +29,37 @@ public class StoredItemController {
         return getId;
     }
 
-    @GetMapping("/getAll")
-    public List<StoredItem> getAllItems(){
-       return storedItemService.getAllItems();
+    @GetMapping("/All")
+    public List<StoredItem> getAllItems() {
+        return storedItemService.getAllItems();
     }
 
-    @GetMapping("/get/{id}")
-    public StoredItemDTO getItem(@PathVariable("id") Long id){
+    @GetMapping("/{id}")
+    public StoredItemDTO getItem(@PathVariable("id") Long id) {
         Optional<StoredItem> storedItemOptional = storedItemService.getItem(id);
-
-        return new StoredItemDTO(storedItemOptional.get());
+        if (storedItemOptional.isPresent()) {
+            return new StoredItemDTO(storedItemOptional.get());
+        }
+        throw new NotFoundException();
     }
 
-    @PutMapping("/update/{id}")
-    public String updateItem(@PathVariable("id") Long id, @RequestParam Long quantity){
-        storedItemService.updateItem(id, quantity);
-        return "item updated";
+    @PostMapping("/update/{id}")
+    public String updateItem(@PathVariable("id") Long id, @RequestParam Long quantity) {
+        Optional<StoredItem> storedItemOptional = storedItemService.getItem(id);
+        if (storedItemOptional.isPresent()) {
+            storedItemService.updateItem(id, quantity);
+            return "item updated";
+        }
+        throw new NotFoundException();
     }
 
-    @DeleteMapping("/delete/{id}")
-    public String deleteItem(@PathVariable("id") Long id){
-        storedItemService.removeItem(id);
-        return "item removed";
+    @DeleteMapping("/{id}")
+    public String deleteItem(@PathVariable("id") Long id) {
+        Optional<StoredItem> storedItemOptional = storedItemService.getItem(id);
+        if (storedItemOptional.isPresent()) {
+            storedItemService.removeItem(id);
+            return "item removed";
+        }
+        throw new NotFoundException();
     }
 }
